### 2025-04-24 #8109 @editorialbot
[PRE REVIEW]: ATMOS-BUD: A Comprehensive Python Tool for Analyzing Heat, Vorticity and Water Budgets on the Atmosphere
**Submitting author:** @daniloceano (<a href="http://orcid.org/0000-0003-4121-7583">Danilo Couto de Souza</a>)
**Repository:** https://github.com/daniloceano/ATMOS-BUD
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** 0.0.0
