### 2025-04-24 #8109 @daniloceano
Hi @kthyng!

I am currently facing a significant issue with the compilation of my paper within the GitHub Actions workflow. The error message provided does not give much information, but from what I could gather, it seems to be related to the Ruby script used for the compilation process, which is beyond my reach:

> from /home/runner/work/_actions/xuanxu/paper-action/main/get_paper.rb:12:in `<main>' Error: Process completed with exit code 1.

However, when attempting to compile locally using several methods, including Pandoc and Inara, I was able to successfully generate the PDF. Below are the commands and configurations I used for local compilation:

**Using Pandoc:**
```bash
pandoc paper/paper.md \
  --from markdown \
  --to latex \
  --citeproc \
  --bibliography=paper/paper.bib \
  -o paper/paper.tex
```

Using Inara:

I cloned the Inara repository and set up the environment and, to ensure the correct recognition of the paper.bib file, I modified the defaults.yaml file in the resources/joss folder to include the line:

`bibliography: <path>/ATMOS-BUD/paper/paper.bib`

Despite these efforts, I have been unable to compile the paper using GitHub Actions. I followed the same structure for the draft-pdf.yml  as other submissions I have made to JOSS, which suggests that the problem is not with it or with the .md file or the .bib file, as they work correctly locally.

Could you please advise if there is any support I can seek or additional steps I can take to overcome this issue?

Thank you for your assistance.
