### 2025-04-25 #8109 @kthyng
@openjournals/dev Can you tell what is wrong with the paper compilation here?
