### 2025-04-25 #8109 @arfon
I am wondering if it might be the structure of the date:

https://github.com/daniloceano/ATMOS-BUD/blob/joss-submission/paper/paper.md?plain=1#L24

@daniloceano – could you try changing this to match the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html), e.g., something like '22 April 2025'
