### 2025-04-26 #8109 @daniloceano
@kthyng  @arfon thank you very much for the assistance. I have changed the date format but it is still getting an error while compiling the PDF (I have also noted an error on the affiliations and fixed that, but following the example paper format so this shouldn't be an issue)
