### 2025-05-05 b355c1d
fixing missing .bib reference on paper
