### 2025-05-05 #8109 @daniloceano
My apologies for missing that! I’ve added `bibliography: paper.bib` in the metadata of the `paper.md` now and the issue is fixed. Thanks so much for pointing that out @xuanxu.

@kthyng  Really appreciate your patience with this!
