+++ paper/paper.md
@@ -32,13 +32,13 @@ bibliography: paper.bib
 
 # Background
 
-The analysis of heat, vorticity, and water budgets is a powerful tool for diagnosing the dynamical and thermodynamical forcings that drive the development of various atmospheric systems [@shu2022vorticity] [@sun2024comparison], particularly cyclones, whether extratropical [@liou1987heat], subtropical [@dutra2017structure], or tropical [@raymond2011vorticity]. The heat budget describes the local temperature tendency as the result of three main processes: horizontal advection, adiabatic heating or cooling due to vertical motion, and diabatic heating processes such as latent heat release. The latter is especially important, as it is associated with convection — a key process that fuels the development of systems such as tropical cyclones. In the vorticity budget, the local change in relative vorticity is governed by several mechanisms: horizontal and vertical advection, advection of planetary vorticity, stretching due to mass divergence or convergence, tilting of horizontal vorticity into the vertical, and frictional effects. Lastly, in the water vapor budget, the local change in column-integrated moisture is primarily controlled by two processes: horizontal moisture flux convergence, and net sources or sinks due to surface evaporation and precipitation.
+The analysis of heat, vorticity, and water budgets is a powerful tool for diagnosing the dynamical and thermodynamical forcings that drive the development of various atmospheric systems [@shu2022vorticity; @sun2024comparison], particularly cyclones, whether extratropical [@liou1987heat], subtropical [@dutra2017structure], or tropical [@raymond2011vorticity]. The heat budget describes the local temperature tendency as the result of three main processes: horizontal advection, adiabatic heating or cooling due to vertical motion, and diabatic heating processes such as latent heat release. The latter is especially important, as it is associated with convection — a key process that fuels the development of systems such as tropical cyclones. In the vorticity budget, the local change in relative vorticity is governed by several mechanisms: horizontal and vertical advection, advection of planetary vorticity, stretching due to mass divergence or convergence, tilting of horizontal vorticity into the vertical, and frictional effects. Lastly, in the water vapor budget, the local change in column-integrated moisture is primarily controlled by two processes: horizontal moisture flux convergence, and net sources or sinks due to surface evaporation and precipitation.
 
 # Statement of need
 
-`ATMOS-BUD` is a Python tool designed to compute heat, vorticity, and water budgets for limited atmospheric regions. It supports both Eulerian and Semi-Lagrangian frameworks. In the former, the user specifies a fixed computational domain for budget computations; in the latter, the domain is updated at each time step to follow the center of an atmospheric system of interest (e.g., a cyclone). This methodology mirrors that adopted in [@de2024lorenzcycletoolkit] (see their Figure 2). Additionally, computations can be performed using an interactive framework, allowing the user to manually define the domain for each time step. The program outputs budget terms averaged over the selected domain and, for moisture variables, provides both layerwise and vertically integrated diagnostics.
+`ATMOS-BUD` is a Python tool designed to compute heat, vorticity, and water budgets for limited atmospheric regions. It supports both Eulerian and Semi-Lagrangian frameworks. In the former, the user specifies a fixed computational domain for budget computations; in the latter, the domain is updated at each time step to follow the center of an atmospheric system of interest (e.g., a cyclone). This methodology mirrors that adopted in @de2024lorenzcycletoolkit (see their Figure 2). Additionally, computations can be performed using an interactive framework, allowing the user to manually define the domain for each time step. The program outputs budget terms averaged over the selected domain and, for moisture variables, provides both layerwise and vertically integrated diagnostics.
 
[…]
