+++ paper/paper.bib
@@ -5,7 +5,8 @@
   volume={115},
   number={9},
   pages={1809--1824},
-  year={1987}
+  year={1987},
+  doi={10.1175/1520-0493(1987)115<1809:HBOAAF>2.0.CO;2}
 }
 
 @article{rew1990netcdf,
@@ -16,7 +17,8 @@
   number={4},
   pages={76--82},
   year={1990},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/38.56302}
 }
 
 
@@ -36,7 +38,8 @@
   number={1},
   pages={147--163},
   year={2011},
-  publisher={Copernicus Publications G{\"o}ttingen, Germany}
+  publisher={Copernicus Publications G{\"o}ttingen, Germany},
+  doi={10.5194/acp-11-147-2011}
 }
 
 @article{dias2013synoptic,
@@ -47,7 +50,8 @@
   number={19},
   pages={10--870},
   year={2013},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/jgrd.50830}
 }
 
 @article{dutra2017structure,
@@ -58,7 +62,8 @@
   number={704},
   pages={1539--1553},
   year={2017},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/qj.3024}
 }
 
 @article{hoyer2017xarray,
@@ -96,7 +101,8 @@
   title={pandas-dev/pandas: Pandas 1.0. 5},
   author={Reback, Jeff and McKinney, Wes and Van Den Bossche, Joris and Augspurger, Tom and Cloud, Phillip and Klein, Adam and Hawkins, Simon and Roeschke, Matthew and Tratner, Jeff and She, Chang and others},
   journal={Zenodo},
-  year={2020}
+  year={2020},
+  doi={10.5281/zenodo.3509134}
 }
 
 
@@ -108,7 +114,8 @@
   number={4},
   pages={556},
   year={2022},
-  publisher={MDPI}
+  publisher={MDPI},
+  doi={10.3390/atmos13040556}
 }
 
 @article{may2022metpy,
@@ -131,7 +138,8 @@
   number={13},
   pages={2396},
   year={2024},
-  publisher={MDPI}
+  publisher={MDPI},
+  doi={10.3390/rs16132396}
 }
 
 @article{de2024lorenzcycletoolkit,
@@ -141,5 +149,6 @@
   volume={9},
   number={101},
   pages={7139},
-  year={2024}
+  year={2024},
+  doi={10.21105/joss.07139}
 }
