### 2025-05-23 d3cf867
Features: min/max tracking and specify vertical level
