+++ src/calculations.py
@@ -6,7 +6,7 @@
 #    By: daniloceano <danilo.oceano@gmail.com>      +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/02/16 16:42:55 by daniloceano       #+#    #+#              #
-#    Updated: 2025/04/21 10:17:23 by daniloceano      ###   ########.fr        #
+#    Updated: 2025/05/23 11:03:03 by daniloceano      ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -143,14 +143,20 @@ def perform_calculations(input_data, namelist_df, dTdt, dZdt, dQdt, args, app_lo
             results_df_dictionary[term] = pd.DataFrame(columns=[str(t) for t in timesteps],
                                             index=[float(i) for i in pres_levels])
         results_nc[term] = []
+
+    # Get choosen vertical level from args and convert to Pa
+    plevel_Pa = int(args.level) * 100
+
+    # Whether to use minimum or maximum vorticity
+    min_max = args.track_vorticity
         
     # Dictionary for saving track attributes for each timestep
     output_track_attributes = {}
     results_keys = ['time', 'central_lat', 'central_lon', 'length', 'width',
-            'min_max_zeta_850','min_hgt_850','max_wind_850']
+            f'{args.track_vorticity}_zeta_{args.level}', f'{args.track_vorticity}_hgt_{args.level}', f'max_wind_{args.level}']
     for key in results_keys:
         output_track_attributes[key] =  []
-    
+
     for time_step in timesteps:
         
         itime = str(time_step)
@@ -171,41 +177,41 @@ def perform_calculations(input_data, namelist_df, dTdt, dZdt, dQdt, args, app_lo
             app_logger=app_logger,
             args=args)
 
-        # Get variables at 850 hPa for the current time step
-        iu_850 = MovingObj.u.sel({vertical_level_indexer:85000})
-        iv_850 = MovingObj.v.sel({vertical_level_indexer:85000})
-        ight_850 = MovingObj.GeopotHeight.sel({vertical_level_indexer:85000})
-        iwspd_850 = wind_speed(iu_850, iv_850)
-        zeta = MovingObj.Zeta.sel({vertical_level_indexer:85000}).metpy.dequantify()
+        # Get variables at choosen pressure level for the current time step
+        iu_plevel = MovingObj.u.sel({vertical_level_indexer:plevel_Pa})
+        iv_plevel = MovingObj.v.sel({vertical_level_indexer:plevel_Pa})
+        ight_plevel = MovingObj.GeopotHeight.sel({vertical_level_indexer:plevel_Pa})
+        iwspd_plevel = wind_speed(iu_plevel, iv_plevel)
+        zeta = MovingObj.Zeta.sel({vertical_level_indexer:plevel_Pa}).metpy.dequantify()
         lat, lon = input_data[MovingObj.latitude_indexer], input_data[MovingObj.longitude_indexer]
 
         # Get domain limits for the current time step
-        variables_at_850hpa  = [iu_850, iv_850, zeta, ight_850, lat, lon, itime]
+        variables_at_plevel  = [iu_plevel, iv_plevel, zeta, ight_plevel, lat, lon, itime]
         if 'track' in locals() or 'track' in globals():
-            current_domain_limits = get_domain_limits(args, *variables_at_850hpa, track=track)
+            current_domain_limits = get_domain_limits(args, *variables_at_plevel, track=track)
         else:
-            current_domain_limits = get_domain_limits(args, *variables_at_850hpa)
[…]
