### 2025-05-23 805ad67
Feature: choose min or max geopotential height
