### 2025-05-23 #8109 @kthyng
@daniloceano
* Is there a reason this isn't available on PyPI or conda-forge?
* Do you have unit tests?
