### 2025-05-28 #8109 @daniloceano
@kthyng I have added the tests :)
