### 2025-06-03 #8109 @kthyng
> As the program is not a package nor was intended to have its functions to be imported independently

@daniloceano If this package is not meant to be imported or used, is it more like an educational tool?
