### 2025-06-03 #8109 @daniloceano
@kthyng

> [@daniloceano](https://github.com/daniloceano) If this package is not meant to be imported or used, is it more like an educational tool?

ATMOS-BUD is intended primarily as a command-line tool, meant to be executed directly from the terminal. It's structured as standalone software rather than a traditional Python package whose functions are individually imported by users, a similar approach to the [LorenzCycleToolKit](https://www.theoj.org/joss-papers/joss.07139/10.21105.joss.07139.pdf).

The software indeed has educational purposes, we've successfully used it in meteorology undergraduate courses at the University of São Paulo to teach students about thermodynamic and vorticity budgets of atmospheric systems. However, the program was also designed with research applications in mind. Researchers and users can easily execute the tool via the command line and directly analyze outputs, without needing to write custom Python scripts.

I currently have a manuscript under review that utilizes ATMOS-BUD, and others are in preparation, so it would be great to be able to cite this work and improve it further based on reviewer suggestions.

I hope this clarifies the intended use of ATMOS-BUD! Let me know if there are any other points you'd like me to address 😄
