### 2025-06-04 #8109 @shenyulu
@daniloceano
**Review of ATMOS-BUD Software Package**

The ATMOS-BUD software suite addresses an important need in atmospheric science education and research by providing integrated tools for calculating heat, vorticity, and moisture balances. While the conceptual framework is commendable, several improvements could enhance its usability and long-term sustainability:

1. **Documentation and Examples**
   The current API documentation, while technically complete, would benefit significantly from:
   - A set of **minimal working examples** (Jupyter notebooks or script examples) demonstrating core functionality
   - Sample input files with standardized formats (e.g., NetCDF or CSV templates) (Although I see samples/sample1_ERA5.nc, but I don't know how to do?)
   - Expected output plots/visualizations with interpretation guidelines
   - A "Quickstart" tutorial showing an end-to-end workflow for common use cases

   *Suggestion*: The educational origins of this package present an excellent opportunity to include pedagogical examples (e.g., case studies from the Synoptic Meteorology course) as part of the documentation.

2. **Testing Infrastructure**
   The absence of unit tests raises concerns about long-term maintainability:
   - Critical calculations (e.g., `CalcZonalAverage`, `CalcAreaAverage`) should have verified test cases against analytical solutions or benchmark datasets
   - A testing framework (pytest/unittest) would facilitate continuous integration (e.g., GitHub Actions)
   - Test coverage should prioritize core numerical routines where floating-point accuracy matters

   *Note*: Meteorological calculations are particularly sensitive to discretization errors—tests should validate both mathematical correctness and numerical stability.

3. **Dependency Management**
   The strict version pinning in requirements.txt may create dependency conflicts:
   - Consider separating "core" vs. "optional" dependencies
   - Version ranges (e.g., `numpy>=1.26.4,<2.0`) could maintain compatibility while allowing updates
   - A conda environment.yml alternative might better handle geoscientific dependencies

   *Technical Note*: After implementing unit tests, the package could adopt a testing approach to validate compatibility across dependency versions (including different python and python packages).

4. **Deployment and Accessibility**
   The current standalone design limits integration potential:
   - PyPI distribution would greatly increase visibility and enable `pip install` workflows
   - Consider add modular function design (independent of GUI/CLI) would allow:
     ```python
     from atmos_bud import vorticity_budget
     budget = vorticity_budget.calculate(u_field, v_field, ...)
     ```

   *Impact*: These changes would align with FAIR (Findable, Accessible, Interoperable, Reusable) principles for research software.

**Additional Recommendations**
- Include contributor guidelines (CONTRIBUTING.md) to encourage community involvement.

The package shows strong potential as both an educational resource and research tool. Implementing these suggestions would significantly increase its adoption and impact in the atmospheric science community.
