### 2025-06-11 43e2ac3
Starting the comprehensive documentation
