+++ docs/index.rst
@@ -14,6 +14,7 @@ Welcome to ATMOS-BUD's documentation!
    getting_started
    downloading_data
    running_the_program
+   fixed_framework_tutorial
    results_and_output
    contact_and_support
    license
