### 2025-06-11 94c8f99
working on saving nc data appropriately
