### 2025-06-12 72944d7
Fixed bug on output nc file and added comprehensive documentation
