+++ plots/plots.py
@@ -1,332 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on Tue Jul 19 09:43:42 2022
-
-@author: danilocoutodsouza
-"""
-
-import matplotlib.pyplot as plt
-import numpy as np
-import pandas as pd
-import matplotlib.dates as mdates
-import cmocean
-import matplotlib.colors as colors
-from matplotlib import cm
-import sys
-import os
-import matplotlib.colors
-
-def check_create_folder(DirName):
-    """
-
-    Check if directory exists and if not, creates it.
-    
-    Parameters
-    ----------
-    DirName : str
-        directory name.
-
-    Returns
-    -------
-    None. 
-
-    """
-    if not os.path.exists(DirName):
-                os.makedirs(DirName)
-                print(DirName+' created')
-    else:
-        print(DirName+' directory exists')
-
-def white_cmap(cmap):
-    n=35
-    x = 0.5
-    lower = cmap(np.linspace(0, x, n))
-    white = np.ones((80-2*n,4))
-    upper = cmap(np.linspace(1-x, 1, n))
-    colors = np.vstack((lower, white, upper))
-    tmap = matplotlib.colors.LinearSegmentedColormap.from_list('map_white', colors)
-    return tmap
-
-
-def time_series(fname,df1,label1,
-                df2=None,label2=None):  
-    """
-
-    Parameters
-    ----------
-    fname : str
-        Name to append to outfile.
-    df1 : pandas.DataFrame
-        Dataframe containing data to be ploted. Indexes are vertical levels and
-        Columns are timesteps
-    label1 : str
-        label for y-axis.
-    df2 : andas.DataFrame, optional
-        same as in df1. The default is None.
-    label2 : str, optional
-        same as in label1. The default is None.
-
-    Returns
-    -------
-    Plot time series for the input data vertical levels closest to 1000, 850, 
-    700, 500, 300 and 200 hPa and save figures. 
-
-    """
-    times = df1.columns
-    pdtime = pd.to_datetime(times) 
-    plt.close('all')
-    fig, axs = plt.subplots(nrows=2, ncols=3, figsize=(14,10))
-    # Find vertical levels that closely matches the desired levels for plotting
-    MatchingLevels = []
-    for pres in [1000,850,700,500,300,200]:
-        MatchingLevels.append(min(df1.index, key=lambda x:abs(x-pres))) 
-    max1,min1 = df1.max().max(),df1.min().min()
-    if df2 is not None:
-        max2,min2 = df2.max().max(),df2.min().min()
-    i = 0
-    for row in range(2):
-        for col in range(3):
-            ax = axs[row,col]
-            p = MatchingLevels[i]
-            ax = axs[row,col]
-            ax.plot(pdtime,df1.loc[p],
-                    color='#BF3D3B',linewidth=2)
-            # If plotting temperature, do not plot horizontal line for 0.
-            if label1 != "T (K)" and df2 is None:
-                ax.axhline(0,zorder=0,c='#383838',alpha=.8,linewidth=.5,
-                            label='')
-            ax.tick_params(axis='x',rotation=45)
-            ax.tick_params(axis='both',which='major',labelsize=10,
-                           labelcolor='#383838')
-            ax.xaxis.set_major_formatter(mdates.DateFormatter('%m-%d'))
-            # If plotting temperature, avoid setting a constant limit for the
-            # y-axis as for each vertical level, the values greatly differ.
-            if label1 != "T (K)":
[…]
