### 2025-06-13 58d8846
Bug fixes in documentation and added interactive framework
