+++ docs/track_framework_tutorial.rst
@@ -41,8 +41,8 @@ The ``track`` file and the ``namelist`` file must be correctly configured in the
 
 Once these steps are completed, you are ready to run the program with the **Track (Semi-Lagrangian) Framework**.
 
-2. Running ATMOS-BUD with the Track Framework
-=============================================
+Running ATMOS-BUD with the Track Framework
+*********************************************
 
 To run ATMOS-BUD using the **Track Framework**, use the following command in your terminal:
 
@@ -130,3 +130,109 @@ After processing each time step, the program will display details like the centr
 
 
 By interpreting this output, users can confirm the successful loading of data, processing of each time step, and generation of output files for further analysis.
+
+Understanding the Output Files
+*********************************************
+
+After running the program, several files and plots are generated. Below is an example of the output directory structure and some of the generated visualizations.
+
+Figures Directory:
+---------------------
+
+  The ``Figures/`` folder contains plots that visualize the defined analysis box and the atmospheric conditions within that box for each time step.
+
+  - ``boxes/``: This subfolder contains images of the defined domain box at different time steps. These images are similar to those in the Fixed Framework and show the region of interest for each time step.
+
+  Example of a figure showing the defined box at 850 hPa for August 8, 2005:
+
+  .. image:: _static/images/box_200508080000_track.png
+     :alt: Box defined for computations
+     :width: 500px
+     :align: center
+
+  - ``track_boxes.png``: This plot shows the track of the system, along with the boxes defined for computations.
+
+  - ``hovmoller_mean_zeta.png``: This is a Hovmöller diagram showing the mean vorticity (`ζ`) over time across the domain.
+
+  - ``timeseries_min_zeta_min_hgt_850hPa.png``: This plot contains the time series of the minimum vorticity (`ζ`) and minimum geopotential height (`hgt`) at 850 hPa (minimum or maximum values of vorticity and geopotential height, depending on the ``--track-vorticity`` and ``--track_geopotential`` arguments). This figure is particularly useful for diagnosing the system's behavior and detecting the phases of its life cycle.
+
+NetCDF File:
+-------------
+ATMOS-BUD will generate a NetCDF file containing all the computed variables for the atmospheric budgets:
+
+  - ``Reg1-Representative_NCEP-R2_track.nc``: This file contains the results of the computations performed on the domain defined by the ``track`` file.
+
+Track File
+-----------
+
+- ``track``: This is the original track file used to define the cyclone’s path.
+
+- ``Reg1-Representative_NCEP-R2_track_track.csv``: This file is generated by the program and includes diagnostic variables for the system's trajectory, computed for the selected pressure level (defined by the ``--level`` argument) and based on the chosen diagnostics (minimum or maximum values of vorticity and geopotential height, depending on the ``--track-vorticity`` and ``--track_geopotential`` arguments). The header of this file looks like:
+
+.. code-block:: bash
+    
+    time;Lat;Lon;length;width;min_zeta_850;min_hgt_850;max_wind_850
+
+
+Where:
+- ``time``: Time step of the track.
+- ``Lat``: Latitude of the system's center.
+- ``Lon``: Longitude of the system's center.
+- ``length`` and ``width``: Dimensions of the box used for the analysis.
+- ``min_zeta_850``: Minimum vorticity at 850 hPa.
+- ``min_hgt_850``: Minimum geopotential height at 850 hPa.
[…]
