### 2025-06-13 da391b5
Updated logging and documentation
