+++ src/output_management.py
@@ -6,7 +6,7 @@
 #    By: daniloceano <danilo.oceano@gmail.com>      +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/02/16 16:13:36 by daniloceano       #+#    #+#              #
-#    Updated: 2025/06/12 08:28:24 by daniloceano      ###   ########.fr        #
+#    Updated: 2025/06/13 10:34:17 by daniloceano      ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -85,9 +85,9 @@ def save_results_netcdf(out_nc, results_subdirectory, outfile_name, app_logger):
     try:
         fname = os.path.join(results_subdirectory, f'{outfile_name}.nc')
         out_nc.to_netcdf(fname, mode='w')
-        app_logger.info(f'{fname} created')
+        app_logger.info(f'💾 {fname} created successfully')
     except Exception as e:
-        app_logger.error(f"Error saving NetCDF file: {e}")
+        app_logger.error(f'❌ Error saving NetCDF file: {e}')
 
 def save_results_csv(results_df_dictionary, results_subdirectory, app_logger):
     """
@@ -118,9 +118,9 @@ def save_results_csv(results_df_dictionary, results_subdirectory, app_logger):
 
             csv_file_name = os.path.join(budget_results_subdirectory, f'{term}.csv')
             df.to_csv(csv_file_name)
-            app_logger.info(f'{csv_file_name} created')
+            app_logger.info(f'💾 {csv_file_name} created successfully')
     except Exception as e:
-        app_logger.error(f"Error saving CSV files: {e}")
+        app_logger.error(f'❌ Error saving CSV files: {e}')
 
 def save_output_track(output_track_attributes, args, results_subdirectory, figures_subdirectory, outfile_name, app_logger):
     """
@@ -139,8 +139,9 @@ def save_output_track(output_track_attributes, args, results_subdirectory, figur
         track = track.rename(columns={'central_lat': 'Lat', 'central_lon': 'Lon'})
         track_file_path = os.path.join(results_subdirectory, f'{outfile_name}_track.csv')
         track.to_csv(track_file_path, index=False, sep=";")
+        app_logger.info(f'💾 {track_file_path} created successfully')
 
         plot_track(track, args, figures_subdirectory, app_logger)
         plot_min_max_zeta_hgt(track.set_index('time'), args, figures_subdirectory, app_logger)
     except Exception as e:
-        app_logger.error(f"Error saving output track: {e}")
+        app_logger.error(f'❌ Error saving output track: {e}')
