+++ src/calculations.py
@@ -6,7 +6,7 @@
 #    By: daniloceano <danilo.oceano@gmail.com>      +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/02/16 16:42:55 by daniloceano       #+#    #+#              #
-#    Updated: 2025/06/12 15:35:17 by daniloceano      ###   ########.fr        #
+#    Updated: 2025/06/13 10:30:11 by daniloceano      ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -160,7 +160,7 @@ def perform_calculations(input_data, namelist_df, dTdt, dZdt, dQdt, args, app_lo
         itime = str(time_step)
         datestr = pd.to_datetime(itime).strftime('%Y-%m-%d %HZ')
         datestr2 = pd.to_datetime(itime).strftime('%Y%m%d%H00')
-        app_logger.info(f'Processing time step: {datestr}')
+        app_logger.info(f'⏳ Processing time step: {datestr}')
 
         # Convert the timezone-aware timestamp 't' to a timezone-naive timestamp (still in UTC)
         t_naive = time_step.tz_localize(None)
@@ -233,24 +233,26 @@ def perform_calculations(input_data, namelist_df, dTdt, dZdt, dQdt, args, app_lo
             NorthernLimit = float(input_data[latitude_indexer].sel({latitude_indexer: max_lat}, method='nearest'))
 
             app_logger.info(
-            f"central lat: {central_lat}, central lon: {central_lon}, "
+            f"📍 central lat: {central_lat}, central lon: {central_lon}, "
             f"size: {length} x {width}, "
             f"lon range: {min_lon} to {max_lon}, "
             f"lat range: {min_lat} to {max_lat}"
             )
 
-        app_logger.info(
-            f"{int(args.level)} hPa diagnostics --> "
+            app_logger.info(
+            f"📍 {int(args.level)} hPa diagnostics --> "
             f"{args.track_vorticity} ζ: {min_max_zeta:.2e}, "
             f"{args.track_geopotential} geopotential height: {min_max_hgt:.0f}, "
             f"max wind speed: {max_wind:.2f}"
-        )
-        app_logger.info(
-            f"{int(args.level)} hPa positions (lat/lon) --> "
+            )
+        
+            app_logger.info(
+            f"📍 {int(args.level)} hPa positions (lat/lon) --> "
             f"{args.track_vorticity} ζ: {min_max_zeta_lat:.2f}, {min_max_zeta_lon:.2f}, "
             f"{args.track_geopotential} geopotential height: {min_max_hgt_lat:.2f}, {min_max_hgt_lon:.2f}, "
             f"max wind speed: {max_wind_lat:.2f}, {max_wind_lon:.2f}")
         
+        app_logger.info(f"📊 Storing results for: {datestr}")
         for term in stored_terms:
             term_sliced = getattr(MovingObj,term).sel(
                 **{
@@ -283,9 +285,12 @@ def perform_calculations(input_data, namelist_df, dTdt, dZdt, dQdt, args, app_lo
             'lat': lat,
             'lon': lon,
         }
+
+        app_logger.info(f"📊 Saving domain plot for {datestr2}")
         plot_fixed_domain(current_domain_limits, dict_for_plot, args, results_subdirectory, datestr2, app_logger)
 
         if args.save_nc_file:
+            app_logger.debug(f"💾 Storing results for the NetCDF file for {datestr}")
             # Initializing out_nc on the first time step
             if time_step == timesteps[0]:
                 # Create the first Dataset with the terms' variables
[…]
