+++ docs/fixed_framework_tutorial.rst
@@ -323,3 +323,20 @@ The following image is an example output generated by this script, showing the t
    :align: center
 
 This visualization clearly highlights regions with positive and negative temperature tendencies, allowing you to quickly interpret the atmospheric dynamics occurring within your defined domain.
+
+Interpreting the Results
+*************************
+
+There is no one-size-fits-all approach to interpreting the results of atmospheric budget calculations for cyclones, as each cyclone has its unique characteristics. These characteristics are reflected in the results of the budgets presented for each system. 
+
+While the figures and CSV files provide valuable diagnostic information, it is important to recognize that cyclonic systems vary significantly in terms of their structure, intensity, and evolution. Therefore, the interpretation of the results requires a case-specific approach. 
+
+To guide the interpretation of the results, there are several studies that describe the physical meaning of each term in the energy, vorticity, and moisture budgets. These articles provide in-depth discussions on the physical processes at play and can serve as valuable references for understanding the results presented by ATMOS-BUD:
+
+1. **Energy cycle and structural evolution of cyclones over southeastern South America**  
+   *Dias Pinto, J. R., and R. P. da Rocha (2011), The energy cycle and structural evolution of cyclones over southeastern South America in three case studies, J. Geophys. Res., 116, D14112*  
+   [Read the article here](https://rmets.onlinelibrary.wiley.com/doi/full/10.1002/qj.3024)
+
+2. **Revisiting The Rare Transition of a South Atlantic Cyclone to Tropical Storm Akara**  
+   *Danilo de Souza et al. (2022), Revisiting The Rare Transition of a South Atlantic Cyclone to Tropical Storm Akara: Energy Cycle and Stratosphere-Troposphere Interaction, Climate Dynamics (Preprint)*  
+   [Read the article here](https://www.researchgate.net/profile/Danilo-De-Souza-3/publication/392520451_Revisiting_The_Rare_Transition_of_a_South_Atlantic_Cyclone_to_Tropical_Storm_Akara_Energy_Cycle_and_Stratosphere-Troposphere_Interaction/links/6846cfdbc33afe388acb0190/Revisiting-The-Rare-Transition-of-a-South-Atlantic-Cyclone-to-Tropical-Storm-Akara-Energy-Cycle-and-Stratosphere-Troposphere-Interaction.pdf)
