### 2025-06-13 0d1485d
implemented benchmarks and respective unit tests
