+++ benchmarks/generate_benchmark.py
@@ -1,0 +1,197 @@
+# benchmarks/generate_benchmarks.py
+
+import xarray as xr
+import numpy as np
+import os
+import sys
+import logging 
+import pandas as pd
+
+from metpy.units import units
+from metpy.calc import vorticity
+
+# Add src to the path
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+from src.calculations import CalcZonalAverage, CalcAreaAverage
+from src.data_object import DataObject
+from src.utils import initialize_logging
+from src.cli_interface import parse_arguments
+from src.data_handling import preprocess_data
+
+def load_and_preprocess_data():
+    """
+    Load the sample dataset and preprocess it by selecting the necessary slices of data
+    and adding the required latitude and longitude coordinates in radians.
+    """
+    # Load the sample dataset
+    ds = xr.open_dataset("./samples/Reg1-Representative_NCEP-R2.nc")
+    
+    # Select the relevant data slice (specific time and level)
+    data = ds['TMP_2_ISBL'].isel(initial_time0_hours=0, lv_ISBL3=-1).isel(lat_2=slice(0, 10), lon_2=slice(0, 10))
+
+    # Add latitude and longitude coordinates in radians and their cosine for zonal averaging
+    data = data.assign_coords({"rlats": np.deg2rad(data['lat_2'])})
+    data = data.assign_coords({"coslats": np.cos(np.deg2rad(data['lat_2']))})
+    data = data.assign_coords({"rlons": np.deg2rad(data['lon_2'])})
+
+    return data
+
+def generate_zonal_average_benchmark():
+    """
+    Generate the benchmark for the CalcZonalAverage function.
+    """
+    data = load_and_preprocess_data()
+
+    # Calculate the zonal average
+    result = CalcZonalAverage(data)
+
+    # Save the result as a benchmark
+    np.save('./benchmarks/benchmark_results/zonal_average.npy', result.values)
+    print("Zonal Average benchmark generated.")
+
+def generate_area_average_benchmark():
+    """
+    Generate the benchmark for the CalcAreaAverage function.
+    """
+    data = load_and_preprocess_data()
+
+    # Calculate the area average (assumed function)
+    result = CalcAreaAverage(data)
+
+    # Save the result as a benchmark
+    np.save('./benchmarks/benchmark_results/area_average.npy', result.values)
+    print("Area Average benchmark generated.")
+
+def calculate_temporal_derivatives(namelist_df, args, app_logger):
+    """
+    Calculate the temporal derivatives (tendencies) for temperature, vorticity, and humidity.
+    """
+    app_logger.info('🔄 Starting the calculation of temporal derivatives for temperature, vorticity, and humidity...')
+    
+    # Load the sample dataset
+    ds = xr.open_dataset("./samples/Reg1-Representative_NCEP-R2.nc")
+
+    data = preprocess_data(ds, namelist_df, args, app_logger)
+    
+    # Select the relevant data slice (specific time and level)
+    data = ds.isel(initial_time0_hours=slice(0, 2), lat_2=slice(0, 10), lon_2=slice(0, 10)).sel(lv_ISBL3=slice(925, 1000))
+
+    # Add latitude and longitude coordinates in radians and their cosine for zonal averaging
+    data = data.assign_coords({"rlats": np.deg2rad(data['lat_2'])})
+    data = data.assign_coords({"coslats": np.cos(np.deg2rad(data['lat_2']))})
+    data = data.assign_coords({"rlons": np.deg2rad(data['lon_2'])})
+
+    # Calculate the temperature tendency (dT/dt)
[…]
