+++ requirements.txt
@@ -1,5 +1,5 @@
-alabaster==0.7.16
-Babel==2.14.0
+alabaster==1.0.0
+babel==2.17.0
 Cartopy==0.22.0
 cdsapi==0.6.1
 certifi==2024.2.2
@@ -11,50 +11,55 @@ cmocean==3.1.3
 contourpy==1.2.0
 cycler==0.12.1
 dask==2024.2.0
-docutils==0.20.1
+docutils==0.21.2
 fonttools==4.49.0
 fsspec==2024.2.0
 idna==3.6
 imagesize==1.4.1
 importlib-metadata==7.0.1
-importlib-resources==6.1.1
-Jinja2==3.1.3
+iniconfig==2.1.0
+Jinja2==3.1.6
 joblib==1.3.2
 kiwisolver==1.4.5
 locket==1.0.0
-MarkupSafe==2.1.5
+lorenz-phase-space==1.2.1
+MarkupSafe==3.0.2
 matplotlib==3.8.3
 MetPy==1.6.1
 netCDF4==1.6.5
 numpy==1.26.4
 packaging==23.2
-pandas==2.2.0
+pandas==2.2.3
 partd==1.4.1
 pillow==10.2.0
 Pint==0.23
 platformdirs==4.2.0
+pluggy==1.6.0
 pooch==1.8.0
 pyarrow==15.0.0
-Pygments==2.17.2
+Pygments==2.19.1
 pyparsing==3.1.1
 pyproj==3.6.1
 pyshp==2.3.1
-python-dateutil==2.8.2
+pytest==8.3.5
+python-dateutil==2.9.0.post0
 pytz==2024.1
 PyYAML==6.0.1
 requests==2.31.0
+roman-numerals-py==3.1.0
 scikit-learn==1.4.1.post1
 scipy==1.12.0
+setuptools==68.2.2
 shapely==2.0.3
 six==1.16.0
-snowballstemmer==2.2.0
-Sphinx==7.2.6
-sphinxcontrib-applehelp==1.0.8
-sphinxcontrib-devhelp==1.0.6
-sphinxcontrib-htmlhelp==2.0.5
+snowballstemmer==3.0.1
+Sphinx==8.2.3
+sphinxcontrib-applehelp==2.0.0
+sphinxcontrib-devhelp==2.0.0
+sphinxcontrib-htmlhelp==2.1.0
 sphinxcontrib-jsmath==1.0.1
-sphinxcontrib-qthelp==1.0.7
-sphinxcontrib-serializinghtml==1.1.10
+sphinxcontrib-qthelp==2.0.0
+sphinxcontrib-serializinghtml==2.0.0
 threadpoolctl==3.3.0
 toolz==0.12.1
 tqdm==4.66.2
@@ -62,5 +67,6 @@ traitlets==5.14.1
 typing_extensions==4.9.0
 tzdata==2024.1
 urllib3==2.2.0
-xarray==2024.1.1
+wheel==0.41.2
+xarray==2025.4.0
 zipp==3.17.0
