+++ .github/workflows/ci.yml
@@ -1,0 +1,31 @@
+name: ATMOS-BUD CI Workflow
+
+on:
+  push:
+    branches:
+      - main
+  pull_request:
+    branches:
+      - main
+
+jobs:
+  test:
+    runs-on: ubuntu-latest  # Usando a última versão do Ubuntu
+
+    steps:
+      - name: Check out code
+        uses: actions/checkout@v2
+
+      - name: Set up Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: '3.10' 
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r requirements.txt  # Instalar dependências de acordo com seu requirements.txt
+
+      - name: Run tests with pytest
+        run: |
+          pytest  # Rodando os testes com pytest
