+++ requirements.txt
@@ -1,4 +1,4 @@
-alabaster==1.0.0
+alabaster==0.7.16
 babel==2.17.0
 Cartopy==0.22.0
 cdsapi==0.6.1
@@ -11,7 +11,8 @@ cmocean==3.1.3
 contourpy==1.2.0
 cycler==0.12.1
 dask==2024.2.0
-docutils==0.21.2
+docutils==0.18.1
+exceptiongroup==1.3.0
 fonttools==4.49.0
 fsspec==2024.2.0
 idna==3.6
@@ -49,11 +50,10 @@ requests==2.31.0
 roman-numerals-py==3.1.0
 scikit-learn==1.4.1.post1
 scipy==1.12.0
-setuptools==68.2.2
 shapely==2.0.3
 six==1.16.0
 snowballstemmer==3.0.1
-Sphinx==8.2.3
+Sphinx==5.0.0
 sphinxcontrib-applehelp==2.0.0
 sphinxcontrib-devhelp==2.0.0
 sphinxcontrib-htmlhelp==2.1.0
@@ -61,12 +61,12 @@ sphinxcontrib-jsmath==1.0.1
 sphinxcontrib-qthelp==2.0.0
 sphinxcontrib-serializinghtml==2.0.0
 threadpoolctl==3.3.0
+tomli==2.2.1
 toolz==0.12.1
 tqdm==4.66.2
 traitlets==5.14.1
 typing_extensions==4.9.0
 tzdata==2024.1
 urllib3==2.2.0
-wheel==0.41.2
 xarray==2025.4.0
 zipp==3.17.0
