### 2025-06-16 #8109 @daniloceano
> **Testing Infrastructure** The absence of unit tests raises concerns about long-term maintainability:
>
> * Critical calculations (e.g., `CalcZonalAverage`, `CalcAreaAverage`) should have verified test cases against analytical solutions or benchmark datasets
> * A testing framework (pytest/unittest) would facilitate continuous integration (e.g., GitHub Actions)
> * Test coverage should prioritize core numerical routines where floating-point accuracy matters

I have addressed the concerns by implementing tests for the core functions, which include all calculation-related functions, such as those for the budgets. I also created benchmark tests where the function results are validated against known results. Specifically, I calculated the budget results for a test dataset, and the script `benchmarks/generate_benchmark.py` was used to generate these benchmark data, ensuring the validity of the calculations.

Additionally, I have integrated GitHub Actions to automate the testing process, so every change will trigger the tests and allow for continuous validation of the code. This setup ensures that any potential issues with floating-point accuracy or other aspects of the core numerical routines are identified early.

I hope these updates address your concerns regarding testing and long-term maintainability. Please let me know if you think further improvements are needed or if you have any additional suggestions! :)
