### 2025-06-19 e1253c1
First attempt to publish to Pypi
