+++ .github/workflows/ci.yml
@@ -19,13 +19,28 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v2
         with:
-          python-version: '3.10' 
+          python-version: '3.10'
 
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt  # Instalar dependências de acordo com seu requirements.txt
 
+      # Etapas para execução dos testes
       - name: Run tests with pytest
         run: |
           pytest  # Rodando os testes com pytest
+
+      # Etapas para criação de distribuição
+      - name: Build distribution
+        run: |
+          python setup.py sdist bdist_wheel  # Criando os arquivos de distribuição
+
+      # Etapas para publicar no PyPI
+      - name: Publish to PyPI
+        run: |
+          pip install twine  # Instala o Twine para o upload
+          twine upload dist/*  # Faz o upload dos pacotes gerados para o PyPI
+        env:
+          PYPI_USERNAME: ${{ secrets.PYPI_USERNAME }}  # Usando o GitHub Secrets
+          PYPI_PASSWORD: ${{ secrets.PYPI_PASSWORD }}  # Usando o GitHub Secrets
