### 2025-06-19 bdf0df7
Added install whell and setuptools to action
