+++ .github/workflows/ci.yml
@@ -47,4 +47,4 @@ jobs:
         env:
           TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}  # Usando o GitHub Secrets para nome de usuário
           TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}  # Usando o GitHub Secrets para senha
-          
+        
