+++ setup.py
@@ -29,8 +29,9 @@ setup(
     },
     classifiers=[
         'Programming Language :: Python :: 3',
-        'License :: OSI Approved :: GNU General Public License v3',
+        'License :: OSI Approved :: GNU General Public License v3.0',
         'Operating System :: OS Independent',
     ],
+    license='GPL-3.0', 
     python_requires='>=3.10',
 )
