### 2025-06-19 912caa1
fixed license classifier again
