+++ setup.py
@@ -29,7 +29,7 @@ setup(
     },
     classifiers=[
         'Programming Language :: Python :: 3',
-        'License :: OSI Approved :: GNU General Public License v3.0',
+        'License :: OSI Approved :: GNU Genereal Public License V3 (GPLV3)',
         'Operating System :: OS Independent',
     ],
     license='GPL-3.0', 
