### 2025-07-24 daniloceano/ATMOS-BUD#6 @crlna16
Upgrade from Python 3.9
Hi! I'm reviewing your JOSS paper here: https://github.com/openjournals/joss-reviews/issues/8452

I noticed the documentation asks to install a conda environment based on Python 3.9. Since this version will not receive security updates after October 2025, I would recommend to bump the version to a later version. The requirements.txt should also be checked for outdated versions.
