### 2025-07-25 daniloceano/ATMOS-BUD#7 @crlna16
New CDS browser: Outdated screenshots, CDS request too large
This is related to the JOSS review: https://github.com/openjournals/joss-reviews/issues/8452

I'm following your tutorial and wanted to download ERA5 data.

The URL is still valid, but I think CDS had an update. The screenshots in the documentation no longer match the actual CDS website, and the order of selecting entries is different. Could you please update the documentation with up-to-date screenshots?

When I enter the request as specified on the documentation, I receive this error:

`Request too large. Cost 64512 over 60000 limit. Please, reduce your selection.`

It would be possible to download it using the CDS API, but apparently not directly from the website. Could you make this explicit in the documentation?
