### 2025-07-25 daniloceano/ATMOS-BUD#8 @crlna16
Add cdsapi to requirements
It is referenced in the documentation and should IMO be part of the requirements.

Related to https://github.com/openjournals/joss-reviews/issues/8452
