### 2025-07-25 daniloceano/ATMOS-BUD#9 @crlna16
Update cdsapi to >=0.7.4
cdsapi is fixed in the requirements to an old version that does not support the new CDS API calls. This is the output I'm getting, using a current API key a request generated from the CDS website:

```
(atmosbud) bash-4.4$ python download.py
/home/k/k202141/.conda/envs/atmosbud/lib/python3.9/site-packages/cdsapi/api.py:17: UserWarning: pkg_resources is deprecated as an API. See https://setuptools.pypa.io/en/latest/pkg_resources.html. The pkg_resources package is slated for removal as early as 2025-11-30. Refrain from using this package or pin to Setuptools<81.
  import pkg_resources
Traceback (most recent call last):
  File "/work/ka1176/caroline/gitlab/ATMOS-BUD/download.py", line 43, in <module>
    client = cdsapi.Client()
  File "/home/k/k202141/.conda/envs/atmosbud/lib/python3.9/site-packages/cdsapi/api.py", line 337, in __init__
    assert len(self.session.auth) == 2, (
AssertionError: The cdsapi key provided is not the correct format, please ensure it conforms to:
<UID>:<APIKEY>
```

Related to: https://github.com/openjournals/joss-reviews/issues/8452
