### 2025-07-28 cd2d90a
Update CDSAPI syntax to latest version in get_era5_data.py
