+++ src/get_era5_data.py
@@ -1,34 +1,45 @@
 import cdsapi
 
-c = cdsapi.Client()
+dataset = "reanalysis-era5-pressure-levels"
+request = {
+    "product_type": ["reanalysis"],
+    "variable": [
+        "geopotential",
+        "specific_humidity",
+        "temperature",
+        "u_component_of_wind",
+        "v_component_of_wind",
+        "vertical_velocity"
+    ],
+    "year": ["2005"],
+    "month": ["08"],
+    "day": [
+        "08", "09", "10",
+        "11", "12", "13",
+        "14"
+    ],
+    "time": [
+        "00:00", "03:00", "06:00",
+        "09:00", "12:00", "15:00",
+        "18:00", "21:00"
+    ],
+    "pressure_level": [
+        "10", "20", "30",
+        "50", "70", "100",
+        "125", "150", "175",
+        "200", "225", "250",
+        "300", "350", "400",
+        "450", "500", "550",
+        "600", "650", "700",
+        "750", "775", "800",
+        "825", "850", "875",
+        "900", "925", "950",
+        "975", "1000"
+    ],
+    "data_format": "netcdf",
+    "download_format": "unarchived",
+    "area": [-17.5, -60, -42.5, -30]
+}
 
-c.retrieve(
-    'reanalysis-era5-pressure-levels',
-    {
-        'product_type':'reanalysis',
-        'format':'netcdf',
-        'pressure_level':[
-            '1','2','3',
-            '5','7','10',
-            '20','30','50',
-            '70','100','125',
-            '150','175','200',
-            '225','250','300',
-            '350','400','450',
-            '500','550','600',
-            '650','700','750',
-            '775','800','825',
-            '850','875','900',
-            '925','950','975',
-            '1000'
-        ],
-        'date':'DATE1/DATE2', # Date format is YYYYMMDD
-        'area':'Nort/West/Sout/East', # Area format is, for example: 90/-180/-90/180
-        'time':'00/to/23/by/1', # "by/1" means downloading hourly data
-        'variable':[
-            'geopotential','vertical_velocity',
-            'temperature','u_component_of_wind',
-            'v_component_of_wind', 'specific_humidity'
-        ]
-    },
-    'outfile.nc') # Output file name
+client = cdsapi.Client()
+client.retrieve(dataset, request).download("system-20050808_ERA5.nc")
