### 2025-07-29 90b6ca0
Fix CI/CD issues and dependency conflicts
- Fix NumPy deprecation warnings in utils.py using .item() method
- Fix Sphinx/docutils dependency conflict by pinning docutils <0.19
- Update cdsapi to version 0.7.6
- Update CHANGELOG with all corrections
