+++ src/get_era5_data.py
@@ -16,12 +16,11 @@ request = {
     "day": [
         "08", "09", "10",
         "11", "12", "13",
-        "14"
+        "14",
     ],
     "time": [
-        "00:00", "03:00", "06:00",
-        "09:00", "12:00", "15:00",
-        "18:00", "21:00"
+        "00:00", "06:00", "12:00",
+        "18:00"
     ],
     "pressure_level": [
         "10", "20", "30",
