+++ requirements.txt
@@ -1,7 +1,5 @@
-alabaster==0.7.16
-babel==2.17.0
 Cartopy==0.22.0
-cdsapi==0.6.1
+cdsapi==0.7.6
 certifi==2024.2.2
 cftime==1.6.3
 charset-normalizer==3.3.2
@@ -11,20 +9,16 @@ cmocean==3.1.3
 contourpy==1.2.0
 cycler==0.12.1
 dask==2024.2.0
-docutils==0.18.1
 exceptiongroup==1.3.0
 fonttools==4.49.0
 fsspec==2024.2.0
 idna==3.6
-imagesize==1.4.1
 importlib-metadata==7.0.1
 iniconfig==2.1.0
-Jinja2==3.1.6
 joblib==1.3.2
 kiwisolver==1.4.5
 locket==1.0.0
 lorenz-phase-space==1.2.1
-MarkupSafe==3.0.2
 matplotlib==3.8.3
 MetPy==1.6.1
 netCDF4==1.6.5
@@ -38,7 +32,6 @@ platformdirs==4.2.0
 pluggy==1.6.0
 pooch==1.8.0
 pyarrow==15.0.0
-Pygments==2.19.1
 pyparsing==3.1.1
 pyproj==3.6.1
 pyshp==2.3.1
@@ -53,13 +46,6 @@ scipy==1.12.0
 shapely==2.0.3
 six==1.16.0
 snowballstemmer==3.0.1
-Sphinx==5.0.0
-sphinxcontrib-applehelp==2.0.0
-sphinxcontrib-devhelp==2.0.0
-sphinxcontrib-htmlhelp==2.1.0
-sphinxcontrib-jsmath==1.0.1
-sphinxcontrib-qthelp==2.0.0
-sphinxcontrib-serializinghtml==2.0.0
 threadpoolctl==3.3.0
 tomli==2.2.1
 toolz==0.12.1
@@ -70,3 +56,23 @@ tzdata==2024.1
 urllib3==2.2.0
 xarray==2025.4.0
 zipp==3.17.0
+
+# Development and documentation dependencies
+alabaster==0.7.16
+babel==2.17.0
+black==23.12.1
+docutils>=0.14,<0.19
+flake8==6.0.0
+imagesize==1.4.1
+Jinja2==3.1.6
+MarkupSafe==3.0.2
+Pygments==2.19.1
+Sphinx==5.0.0
+sphinx-rtd-theme==1.3.0
+sphinxcontrib-applehelp==2.0.0
+sphinxcontrib-devhelp==2.0.0
+sphinxcontrib-htmlhelp==2.1.0
+sphinxcontrib-jsmath==1.0.1
+sphinxcontrib-qthelp==2.0.0
+sphinxcontrib-serializinghtml==2.0.0
+twine==4.0.2
