+++ CHANGELOG.md
@@ -13,13 +13,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated CDSAPI syntax in `get_era5_data.py` to support latest API version
 - Added troubleshooting documentation for "request too large" errors
 - Added CDO merge command example for combining multiple NetCDF files
+- Added `sphinx-rtd-theme` for improved documentation rendering
+- Added development dependencies: `flake8`, `black`, and `twine`
 
 ### Changed
 - Updated data download script to use modern CDSAPI syntax
 - Improved documentation with additional error handling guidance
+- Enhanced requirements.txt with documentation and development tools
+- Updated `cdsapi` from 0.6.1 to 0.7.6
 
 ### Fixed
 - Fixed deprecated CDSAPI syntax that was causing compatibility issues
+- Fixed NumPy deprecation warnings in `src/utils.py` by using `.item()` method for scalar conversion
+- Fixed Sphinx/docutils dependency conflict by pinning docutils to compatible version (<0.19)
 
 ## [0.0.1] - 2025-06-19
 
