### 2025-07-29 daniloceano/ATMOS-BUD#6 @daniloceano
Hi @crln16,

Thanks for the feedback! This has been resolved in our recent updates:

✅ **Python requirement updated to 3.10+** (from 3.9)
✅ **Dependencies updated** including `cdsapi` to latest version (0.7.6)
✅ **Documentation updated** to reflect Python 3.10+ installation

The security concern is now addressed. All changes are available in v0.1.0.

Thanks for the review!
