### 2025-07-29 daniloceano/ATMOS-BUD#7 @daniloceano
Hi @crlna16,

Thanks for the detailed feedback! I've completely updated the data downloading documentation to address these issues:

✅ **Removed outdated screenshots** - Made documentation interface-agnostic to avoid future invalidation
✅ **Updated CDSAPI script** to latest syntax (v0.7.6)
✅ **Added NCAR RDA alternative** as additional download option
✅ **Added a workaround** for "Request too large" errors with CDO for daily downloads

The "Request too large" error is indeed a Copernicus limitation. The documentation now explains how to use CDO to download daily files and merge them. Additionally NCAR RDA may not have the same restrictions, so you might want to try it.

All updates are available in v0.1.0. Thanks for helping improve
