### 2025-07-29 daniloceano/ATMOS-BUD#9 @daniloceano
Hi @crlna16,

Thanks for reporting this! The cdsapi version has been updated to **0.7.6** (latest) in our recent release to support the new CDS API format.

✅ **Updated cdsapi to v0.7.6** - Now supports new API key format and latest CDS calls
✅ **Updated download script syntax** - Compatible with current CDS API

Please update your environment with the latest requirements.txt from v0.1.0 to resolve this issue.

Thanks.
