### 2025-07-29 #8452 @daniloceano
Hi @crlna16,

Thank you for starting the review and for the positive feedback about the package!

I'm pleased to inform you that all the issues you filed have been addressed in our recent updates (v0.1.0):

✅ **CDS/cdsapi issues resolved** - Updated to latest version with new API support
✅ **Python version updated** to 3.10+ for security
✅ **Documentation updated** with current download methods and troubleshooting
✅ **Requirements fixed** - All dependencies updated and conflicts resolved

The installation and environment setup should now work smoothly. Please feel free to **resume your review** - you should be able to reproduce the examples without issues now.

Thanks for your patience and thorough review process!
