### 2025-08-01 daniloceano/ATMOS-BUD#11 @crlna16
Minor: Update 3.9 -> 3.10 in the documentation
Thanks for the quick update! Installation worked for me, and I was able to download the CDS demo data.

Minor correction: in the Install page, the conda environment still is created with python 3.9 instead of python 3.10

https://atmos-bud.readthedocs.io/en/latest/getting_started.html

Relates to JOSS review: https://github.com/openjournals/joss-reviews/issues/8452
