### 2025-08-01 daniloceano/ATMOS-BUD#12 @crlna16
Fixed framework tutorial: figures -> plots
I was able to execute the fixed framework tutorial. Minor comment: The commands should be changed from e.g.

`python figures/map_example.py`

to

`python plots/map_example.py`

Relates to: https://github.com/openjournals/joss-reviews/issues/8452
