### 2025-08-01 daniloceano/ATMOS-BUD#13 @crlna16
Add API documentation
I might have missed it, but I did not see an API documentation for the functions in `src`. Could you add that to the documentation? Docstrings etc. look good, so it should be possible to generate the API documentation in an automated way.

Relates to: https://github.com/openjournals/joss-reviews/issues/8452
