### 2025-08-01 #8452 @crlna16
Hello @daniloceano ! I was able to install the package and raised some more issues regarding documentation. Everything worked as described.

The usability and maintainability of the package is in line with typical academic software that is developed by a small group of people and with specific users in mind. You may want to consider three things:
- Make it installable as a package (by adding setup.py or pyproject.toml) instead of the conda -> pip install -> execute python script logic
- Add Gitlab CI/CD pipeline that automatically executes the tests and reports coverage to help maintain code quality
- Add Jupyter notebooks that illustrate standard use cases, this can be tremendously helpful to onboard new users

However, given the target audience of this package, I would be fine even without these points.

@AnjaliSandip I'll be out of office until the end of August, but once the issues are closed by @daniloceano, I would consider my review finished.
