### 2025-08-15 e202373
Version 0.1.1: Added API documentation and improved test coverage
