+++ src/get_era5_data.py
@@ -1,44 +1,243 @@
-import cdsapi
+"""
+ERA5 Data Download Module
 
-dataset = "reanalysis-era5-pressure-levels"
-request = {
-    "product_type": ["reanalysis"],
-    "variable": [
+This module provides functionality for downloading ERA5 atmospheric reanalysis data 
+from the Copernicus Climate Data Store (CDS). It handles authentication, data requests, 
+and automatic file management for ATMOS-BUD workflows.
+
+Author: Danilo Couto de Souza
+Date: 2024
+"""
+
+import cdsapi
+import os
+import logging
+from datetime import datetime, timedelta
+from typing import List, Optional, Union
+
+
+def download_era5_data(
+    variables: List[str],
+    pressure_levels: List[int],
+    start_date: str,
+    end_date: str,
+    area: List[float],
+    output_file: str,
+    hours: Optional[List[str]] = None,
+    logger: Optional[logging.Logger] = None
+) -> None:
+    """
+    Downloads ERA5 reanalysis data from the Copernicus Climate Data Store using the modern CDSAPI.
+    
+    This function provides a flexible interface for downloading ERA5 atmospheric data
+    with configurable variables, pressure levels, time periods, and spatial domains.
+    Uses the modern CDSAPI 0.7.6+ syntax with list-based parameters.
+    
+    Parameters
+    ----------
+    variables : List[str]
+        List of ERA5 variable names to download. Examples:
+        - 'temperature'
+        - 'u_component_of_wind' 
+        - 'v_component_of_wind'
+        - 'geopotential'
+        - 'vorticity'
+        - 'specific_humidity'
+    
+    pressure_levels : List[int]
+        List of pressure levels in hPa. Examples: [1000, 925, 850, 700, 500, 300]
+    
+    start_date : str
+        Start date in 'YYYY-MM-DD' format (e.g., '2023-01-01')
+    
+    end_date : str  
+        End date in 'YYYY-MM-DD' format (e.g., '2023-01-31')
+    
+    area : List[float]
+        Spatial domain as [North, West, South, East] in decimal degrees.
+        Example: [20, -80, -60, -20] for South America region
+    
+    output_file : str
+        Output filename for the downloaded NetCDF file
+        
+    hours : Optional[List[str]], default None
+        List of hours in 'HH:MM' format. If None, uses ['00:00', '06:00', '12:00', '18:00']
+        
+    logger : Optional[logging.Logger], default None
+        Logger object for progress tracking and error reporting
+        
+    Returns
+    -------
+    None
+        Downloads file directly to specified output path
+        
+    Raises
+    ------
+    Exception
+        If download fails due to authentication, network, or API errors
+        
+    Notes
+    -----
+    - Requires valid CDS API credentials in ~/.cdsapirc
+    - Uses modern CDSAPI syntax (version 0.7.6+)
+    - Automatically handles date range conversion to required format
+    - Includes progress monitoring and comprehensive error handling
+    
+    Examples
+    --------
+    >>> # Basic usage for atmospheric budget analysis
+    >>> variables = ['temperature', 'u_component_of_wind', 'v_component_of_wind', 'geopotential']
+    >>> levels = [850, 700, 500, 300]
+    >>> download_era5_data(
+    ...     variables=variables,
+    ...     pressure_levels=levels,
+    ...     start_date='2023-01-01',
+    ...     end_date='2023-01-31', 
+    ...     area=[10, -80, -40, -30],
+    ...     output_file='era5_january2023.nc'
+    ... )
+    
+    >>> # Custom time selection
+    >>> download_era5_data(
+    ...     variables=['vorticity', 'geopotential'],
+    ...     pressure_levels=[850],
+    ...     start_date='2023-06-15',
+    ...     end_date='2023-06-15',
[…]
