### 2025-08-15 #8452 @daniloceano
Hi @crlna16,

Thank you for the comprehensive review and valuable feedback! I've addressed all the key points:

### ✅ **Package Structure & CI/CD**
The project is now properly configured as a Python package with automated CI/CD pipeline for testing. The package can be installed via pip and includes all necessary dependencies and configuration files.

### 📚 **Jupyter Notebooks vs Documentation**
Regarding Jupyter notebooks, I appreciate the suggestion, but given ATMOS-BUD's primary design as a **command-line tool** rather than an interactive library, traditional documentation tutorials are more appropriate for our use case. Users typically execute the tool via terminal commands (`python main.py`) rather than importing modules interactively. The comprehensive documentation with step-by-step tutorials serves this workflow better than notebooks would.

Plus: All suggested changes in the API and documentation are available in the latest release. Thank you for helping improve the project quality!
