### 2025-08-15 1feb264
Bump version to 0.1.2 to resolve PyPI upload conflict
- Fix CI/CD pipeline issue where version 0.1.1 already exists on PyPI
- Update version in setup.py, src/__init__.py, and CHANGELOG.md
- This allows successful package publication to PyPI
