+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
- 
+## [0.1.2] - 2025-08-15
+
+### Fixed
+- Resolved PyPI upload conflict by incrementing version number
+- Fixed CI/CD pipeline issues with package publishing
+
 ## [0.1.1] - 2025-08-15
 
 ### Added
