### 2025-08-15 bc5002c
Bump version to 0.1.3 - Fix duplicate CI/CD PyPI upload
- Resolved issue where CI/CD runs on both develop push AND main merge
- This prevents double publication attempts to PyPI
- Future fix: Configure workflow to publish only on main branch
