### 2025-08-15 9fc5b12
Fix CI/CD: Prevent duplicate PyPI uploads during PRs
- Add conditional to publish only on push to main, not on PRs
- Prevents double publication when PR is created AND merged
- Maintains testing on both PRs and pushes for validation
- This resolves the 'file already exists' PyPI upload errors
