+++ .github/workflows/ci.yml
@@ -10,7 +10,7 @@ on:
 
 jobs:
   test:
-    runs-on: ubuntu-latest  # Usando a última versão do Ubuntu
+    runs-on: ubuntu-latest
 
     steps:
       - name: Check out code
@@ -24,27 +24,28 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt  # Instalar dependências de acordo com seu requirements.txt
+          pip install -r requirements.txt
 
       - name: Install setuptools and wheel
         run: |
-          pip install setuptools wheel  # Instalar setuptools e wheel para a criação de pacotes
+          pip install setuptools wheel
 
       - name: Run tests with pytest
         run: |
-          pytest  # Rodando os testes com pytest
+          pytest
 
-      # Etapas para criação de distribuição
+      # Build distribution (sempre executa para validar)
       - name: Build distribution
         run: |
-          python setup.py sdist bdist_wheel  # Criando os arquivos de distribuição
+          python setup.py sdist bdist_wheel
 
-      # Etapas para publicar no PyPI
+      # Publish to PyPI (APENAS quando push para main, não em PRs)
       - name: Publish to PyPI
+        if: github.event_name == 'push' && github.ref == 'refs/heads/main'
         run: |
-          pip install twine  # Instala o Twine para o upload
-          twine upload dist/*  # Faz o upload dos pacotes gerados para o PyPI
+          pip install twine
+          twine upload dist/*
         env:
-          TWINE_USERNAME: "__token__"  # Nome de usuário no PyPI para autenticação com token
-          TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}  # Usando o GitHub Secrets para o API token
+          TWINE_USERNAME: "__token__"
+          TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
         
