### 2025-08-15 295723e
Release v0.1.4: Fix ReadTheDocs documentation build
### Fixed
- Resolved ReadTheDocs build failures by switching to simplified Sphinx configuration
- Fixed documentation compatibility issues in RTD environment
- Enhanced ReadTheDocs integration with proper configuration files

### Added
- ReadTheDocs configuration file (.readthedocs.yaml) for consistent builds
- Simplified conf.py for better RTD compatibility

### Updated
- Version bumped to 0.1.4 across setup.py, __init__.py, and CHANGELOG.md
- Documentation should now build correctly on ReadTheDocs
