+++ src/__init__.py
@@ -16,10 +16,10 @@ ATMOS-BUD: Atmospheric Budget Analysis Tool
 A comprehensive Python package for analyzing heat, vorticity, and humidity budgets
 of limited regions in the atmosphere using reanalysis data.
 
-Version: 0.1.3
+Version: 0.1.4
 """
 
-__version__ = "0.1.3"
+__version__ = "0.1.4"
 __author__ = "Danilo Couto de Souza"
 __email__ = "danilo.oceano@gmail.com"
 __license__ = "GPL-3.0"
