+++ docs/requirements.txt
@@ -1,1 +1,3 @@
 sphinx==7.2.6
+sphinx-rtd-theme==2.0.0
+docutils<0.19
