+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.4] - 2025-08-15
+
+### Fixed
+- Fixed ReadTheDocs build issues by simplifying Sphinx configuration
+- Resolved documentation build failures in ReadTheDocs environment
+- Switched to simplified `conf.py` for better compatibility
+- Enhanced ReadTheDocs integration with proper `.readthedocs.yaml` configuration
+
+### Added
+- ReadTheDocs configuration file (`.readthedocs.yaml`) for consistent builds
+- Simplified Sphinx configuration for better RTD compatibility
+
 ## [0.1.3] - 2025-08-15
 
 ### Fixed
