### 2025-08-15 741c3d4
docs: Fix version number in Sphinx configuration
Update docs/conf.py to show correct version 0.1.4 in ReadTheDocs
documentation. This is a documentation-only fix that doesn't
require a new package version.
