### 2025-08-15 15a7935
Release v0.1.5: Fix ReadTheDocs version display and CI/CD pipeline
- Update version to 0.1.5 in setup.py, src/__init__.py, and docs/conf.py
- Fix ReadTheDocs documentation showing incorrect version number
- Resolve CI/CD pipeline failure caused by duplicate PyPI publication
- Ensure consistent version numbers across all configuration files

This patch release addresses infrastructure issues without code changes.
