### 2025-09-01 #8452 @shenyulu
I thank the authors for improving the manuscript and repo. All my concerns have been fully addressed. I recommend publishing it in joss.
