### 2025-09-22 771141d
Release v0.1.6: Update version numbers and improve documentation consistency
