### 2025-10-11 b30ec3d
Fix capitalization in titles of articles in paper.bib for consistency
