+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{liou1987heat,
-  title={Heat budgets of analyses and forecasts of an explosively deepening maritime cyclone},
+  title={{Heat Budgets of Analyses and Forecasts of an Explosively Deepening Maritime Cyclone}},
   author={Liou, Chi-Sann and Elsberry, Russel L},
   journal={Monthly weather review},
   volume={115},
@@ -10,7 +10,7 @@
 }
 
 @article{rew1990netcdf,
-  title={NetCDF: an interface for scientific data access},
+  title={{NetCDF: An Interface for Scientific Data Access}},
   author={Rew, Russ and Davis, Glenn},
   journal={IEEE computer graphics and applications},
   volume={10},
@@ -23,7 +23,7 @@
 
 
 @article{doty1995geophysical,
-  title={Geophysical data analysis and visualization using the grid analysis and display system},
+  title={{Geophysical Data Analysis and Visualization Using the Grid Analysis and Display System}},
   author={Doty, Brian E and Kinter, JLI},
   journal={National Aeronautics and Space Administration, Washington, DC, USA},
   year={1995}
@@ -31,7 +31,7 @@
 
 
 @article{raymond2011vorticity,
-  title={The vorticity budget of developing typhoon Nuri (2008)},
+  title={The Vorticity Budget of Developing Typhoon Nuri (2008)},
   author={Raymond, DJ and L{\'o}pez Carrillo, C},
   journal={Atmospheric Chemistry and Physics},
   volume={11},
@@ -43,7 +43,7 @@
 }
 
 @article{dias2013synoptic,
-  title={Synoptic and dynamical analysis of subtropical cyclone Anita (2010) and its potential for tropical transition over the South Atlantic Ocean},
+  title={{Synoptic and Dynamical Analysis of Subtropical Cyclone Anita (2010) and Its Potential for Tropical Transition over the South Atlantic Ocean}},
   author={Dias Pinto, Joao Rafael and Reboita, Michelle Sim{\~o}es and da Rocha, Rosmeri Porf{\'\i}rio},
   journal={Journal of Geophysical Research: Atmospheres},
   volume={118},
@@ -55,7 +55,7 @@
 }
 
 @article{dutra2017structure,
-  title={Structure and evolution of subtropical cyclone Anita as evaluated by heat and vorticity budgets},
+  title={{Structure and Evolution of Subtropical Cyclone Anita as Evaluated by Heat and Vorticity Budgets}},
   author={Dutra, L{\'\i}via M{\'a}rcia Mosso and da Rocha, Rosmeri Porf{\'\i}rio and Lee, Robert William and Peres, Jean Rafael Rom{\~a}o and de Camargo, Ricardo},
   journal={Quarterly Journal of the Royal Meteorological Society},
   volume={143},
@@ -67,7 +67,7 @@
 }
 
 @article{hoyer2017xarray,
-  title={{xarray: ND Labeled Arrays and Datasets in Python}},
+  title={{xarray: N-D Labeled Arrays and Datasets in Python}},
   author={Hoyer, Stephan and Hamman, Joe},
   journal={Journal of Open Research Software},
   volume={5},
@@ -98,7 +98,7 @@
 }
 
 @article{reback2020pandas,
-  title={pandas-dev/pandas: Pandas 1.0. 5},
+  title={{pandas-dev/pandas: Pandas 1.0.5}},
   author={Reback, Jeff and McKinney, Wes and Van Den Bossche, Joris and Augspurger, Tom and Cloud, Phillip and Klein, Adam and Hawkins, Simon and Roeschke, Matthew and Tratner, Jeff and She, Chang and others},
[…]
