### 2025-10-12 f30e799
Fix title formatting for consistency in paper.bib
