+++ paper/paper.bib
@@ -31,7 +31,7 @@
 
 
 @article{raymond2011vorticity,
-  title={The Vorticity Budget of Developing Typhoon Nuri (2008)},
+  title={{The Vorticity Budget of Developing Typhoon Nuri (2008)}},
   author={Raymond, DJ and L{\'o}pez Carrillo, C},
   journal={Atmospheric Chemistry and Physics},
   volume={11},
